+++ src/features/packages/flows/pav/PAVFlow.tsx
@@ -410,8 +410,8 @@ export default function PAVFlow({ onExitFlow }: { onExitFlow?: () => void }) {
                     eventCategories: eventCategories ? eventCategories.split(',').map(s => s.trim()) : ['Wedding'],
                     poc: pocName,
                     duration: {
-                        minHours: parseInt(minDuration) || 0,
-                        maxHours: parseInt(maxDuration) || 0,
+                        minHours: parseFloat(minDuration) || 0,
+                        maxHours: parseFloat(maxDuration) || 0,
                     },
                     crewSize: {
                         minPeople: totalCrewSize,
